started splitting ewcfg
